transcriber_project/requirements.txt
openai-whisper
pydub
ffmpeg-python
